admin_panel/requirements.txt
Django==4.2.8
python-dotenv==1.0.0
Pillow==10.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
